+++ tests/web_tests/test_element.py
@@ -91,9 +91,24 @@ def test_screenshot_remove(base_playground_page):
 # Test waits
 
 
+def test_wait_element(pizza_order_page):
+    pizza_order_page.submit_button.wait_element()
+    assert pizza_order_page.submit_button.is_displayed()
+
+
 def test_wait_without_error(pizza_order_page):
-    pizza_order_page.error_modal.wait_element_without_error(timeout=0.01)
+    pizza_order_page.error_modal.wait_element_without_error(timeout=0.5)
     assert not pizza_order_page.error_modal.is_displayed()
+
+
+def test_wait_hidden(pizza_order_page):
+    pizza_order_page.error_modal.wait_element_without_error()
+    assert not pizza_order_page.error_modal.is_displayed()
+
+
+def test_wait_hidden_without_error(pizza_order_page):
+    pizza_order_page.submit_button.wait_element_without_error(timeout=0.5)
+    assert pizza_order_page.submit_button.is_displayed()
 
 
 def test_click_and_wait(pizza_order_page, driver_engine):